data.sql, new cruds, h2 database configuration, some minor changes

+++ src/main/java/com/demo/springboot/assignment_three/services/InstructorService.java
@@ -29,8 +29,14 @@ public class InstructorService {
         instructorRepository.save(instructor);
     }
 
-    public void deleteById(Long id){
-        instructorRepository.deleteById(id);
+    public boolean deleteById(Long id){
+        Optional<Instructor> instructorOptional = instructorRepository.findById(id);
+        if (instructorOptional.isPresent()) {
+            instructorRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
